Refactor plotting utilities
