+++ run_pipeline.py
@@ -127,7 +127,7 @@ def generate_timeline_chart(timings):
         str: Ruta al archivo del gráfico generado
     """
     try:
-        from sp500_analysis.shared.visualization.plotters import plot_pipeline_timeline
+        from sp500_analysis.shared.visualization.time_series_plots import plot_pipeline_timeline
 
         chart_path = REPORTS_DIR / "pipeline_timeline.png"
         plot_pipeline_timeline(timings, chart_path)
